+++ libs/effects/src/lib/effects-manager.ts
@@ -1,13 +1,9 @@
-import { Observable, Subject, Subscription, map } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { Actions, actions } from './actions';
 import { Action } from './actions.types';
-import { Effect, EffectConfig } from './effects.types';
+import { Effect } from './effects.types';
 import { coerceArray } from './utils';
-import {
-  ActionCreatorIsNotAllowed,
-  ActionCreatorsAreNotAllowed,
-} from './action-creator-is-not-allowed.type';
 
 export interface EffectsConfig {
   dispatchByDefault?: boolean;
@@ -95,33 +91,6 @@ export function initEffects(config?: EffectsConfig) {
 
   return (effectsManager = new EffectsManager(config));
 }
-
-export function createEffect<T>(
-  factory: (actions: Actions) => Observable<T>,
-  config?: { dispatch: false }
-): Effect<T>;
-export function createEffect<T extends Action>(
-  factory: (actions: Actions) => Observable<ActionCreatorIsNotAllowed<T>>,
-  config: { dispatch: true }
-): Effect<ActionCreatorIsNotAllowed<T>>;
-export function createEffect<T extends Action>(
-  factory: (actions: Actions) => Observable<ActionCreatorIsNotAllowed<T>[]>,
-  config: { dispatch: true }
-): Effect<ActionCreatorIsNotAllowed<T>[]>;
-export function createEffect<T extends Action>(
-  factory: (
-    actions: Actions
-  ) => Observable<
-    ActionCreatorIsNotAllowed<T> | ActionCreatorsAreNotAllowed<T[]> | any
-  >,
-  config?: EffectConfig
-): Effect {
-  return { sourceFn: factory, config };
-}
-
-createEffect((actions) => actions.pipe(map(() => [{ type: '' }])), {
-  dispatch: true,
-});
 
 export function registerEffects(effects: Effect | Effect[]) {
   effectsManager.registerEffects(coerceArray(effects));
